Adding both implementations of geodesics
