+++ addon/operator/measures_geodesic_operator.py
@@ -6,6 +6,7 @@ import traceback
 from ..utility.draw import draw_messages
 from ..utility.ray import mouse_raycast_to_scene
 from .geopath_datastructure import GeoPath, Geodesic_State
+# from .geopath_datastructure_vertices import GeoPath, Geodesic_State
 
 
 class MEASURES_GEODESIC_OT(bpy.types.Operator):
